병합
병함

+++ back/src/main/java/PHCCS/web/controller/CommentController.java
@@ -1,6 +1,7 @@
 package PHCCS.web.controller;
 
 import PHCCS.domain.Comment;
+import PHCCS.jwt.JwtUtil;
 import PHCCS.web.repository.domain.CommentDTO;
 import PHCCS.web.service.CommentService;
 import PHCCS.web.service.SSEService;
@@ -18,20 +19,23 @@ import java.util.List;
 @RequiredArgsConstructor
 public class CommentController {
 
+    private final JwtUtil jwtUtil;
     private final CommentService service;
     private final SSEService sseService;
 
     @PostMapping("/add/{category}/{id}")
     public ResponseEntity<?> addComment(
 //            @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member loginMember,
+            @RequestHeader("Authorization") String token,
             @PathVariable("category") String category,
             @PathVariable("id") Long postId,
             @RequestBody Comment comment){
 
         log.info("postComment()");
-//        if(!isLogin(loginMember)){
-//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
-//        }
+        Long loginMember = jwtUtil.extractSubject(token);
+        if(loginMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
+        }
         boolean isSave = service.save(category, postId, comment);
         if(isSave){
             sseService.addCommentAlarm(category, postId, comment);
@@ -41,15 +45,17 @@ public class CommentController {
     }
 
     @GetMapping("/show/{category}/{id}")
-    public ResponseEntity<List<Comment>> findAllComment(
+    public ResponseEntity<?> findAllComment(
 //            @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member loginMember,
+            @RequestHeader("Authorization") String token,
             @PathVariable("category") String category,
             @PathVariable("id") Long postId){
 
         log.info("findAllComment()");
-//        if(!isLogin(loginMember)){
-//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
-//        }
+        Long loginMember = jwtUtil.extractSubject(token);
+        if(loginMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
+        }
         List<Comment> allComment = service.findAllComment(category, postId);
         return ResponseEntity.ok(allComment);
     }
@@ -58,14 +64,16 @@ public class CommentController {
     @PutMapping("/update/{category}/{postId}/{commentId}")
     public ResponseEntity<?> updateComment(
 //            @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member loginMember,
+            @RequestHeader("Authorization") String token,
             @PathVariable("category") String category,
             @PathVariable("postId") Long postId,
             @PathVariable("commentId") Long commentId,
             @RequestBody CommentDTO dto){
         log.info("updateComment()");
-//        if(!isLogin(loginMember)){
-//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
-//        }
+        Long loginMember = jwtUtil.extractSubject(token);
+        if(loginMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
+        }
         service.updateComment(category, postId, commentId, dto);
 
         return ResponseEntity.ok(HttpStatus.OK);
@@ -74,14 +82,16 @@ public class CommentController {
     @DeleteMapping("delete/{category}/{postId}/{commentId}")
     public ResponseEntity<?> deleteComment(
 //            @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member loginMember,
+            @RequestHeader("Authorization") String token,
             @PathVariable("category") String category,
             @PathVariable("postId") Long postId,
             @PathVariable("commentId") Long commentId){
 
         log.info("deleteComment()");
-//        if(!isLogin(loginMember)){
-//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
-//        }
+        Long loginMember = jwtUtil.extractSubject(token);
+        if(loginMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
+        }
         service.deleteComment(category, postId, commentId);
         return null;
     }
@@ -89,13 +99,15 @@ public class CommentController {
     @PostMapping("/like/{category}/{postId}/{commentId}")
     public ResponseEntity<?> incrementLike(
 //            @SessionAttribute(name = SessionConst.LOGIN_MEMBER, required = false) Member loginMember,
+            @RequestHeader("Authorization") String token,
             @PathVariable("category") String category,
             @PathVariable("postId") Long postId,
             @PathVariable("commentId")Long commentId){
         log.info("incrementLike()");
-//        if(!isLogin(loginMember)){
-//            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
-//        }
+        Long loginMember = jwtUtil.extractSubject(token);
+        if(loginMember == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인하지 않은 사용자는 접근할 수 없습니다.");
+        }
         service.incrementLike(/*loginMember.getId()*/2L, category, postId, commentId);
         return ResponseEntity.ok(HttpStatus.OK);
     }
